requirements.txt
amqp
anyio
asgiref
billiard
celery
Cerberus
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
crispy-bootstrap5
decorator
distlib
dj-database-url
dj-rest-auth
Django
django-allauth
django-bootstrap5
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-email
django-environ
django-filter
django-modeltranslation
django-rosetta
django-taggit
djangorestframework
djangorestframework_simplejwt
drf-yasg
Faker
filelock
h11
httpcore
httpx
idna
inflection
kolo
kombu
Markdown
mocket
more-itertools
MouseInfo
msgpack
packaging
pillow
pipenv
platformdirs
polib
prompt_toolkit
psycopg2-binary
puremagic
PyAutoGUI
pycryptodome
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
python-dateutil
python-dotenv
python-utils
pytweening
pytz
PyYAML
requests
setuptools
six
sniffio
sqlglot
sqlparse
stripe
tabulate
toolz
typing_extensions
tzdata
ulid-py
uritemplate
urllib3
vine
virtualenv
virtualenvwrapper-win
wcwidth
whitenoise
